Fixes #257 java-codegen: Reorganize Java Codegen ITs
