+++ language-support/java/codegen/BUILD.bazel
@@ -15,7 +15,8 @@ load(
 
 load(
     ":codegen.bzl",
-    "daml_to_java"
+    "daml_to_java",
+    "mangle_for_java"
 )
 
 load(
@@ -97,12 +98,21 @@ daml_lf_target_versions = ["1.0", "1.1", "1.3"]
 ####  Integration Tests
 ########################################################
 [
-    daml_to_java(
-        name = "integration-tests-model-%s" % target,
-        main_src = "src/it/daml/Lib.daml",
-        srcs = glob(["src/it/daml/**/*.daml"]),
-        target = target,
-    )
+    [
+        daml_to_java(
+            name = "integration-tests-model-noprefix-%s" % target,
+            main_src = "src/it/daml/Lib.daml",
+            srcs = glob(["src/it/daml/**/*.daml"]),
+            target = target,
+        ),
+        daml_to_java(
+            name = "integration-tests-model-%s" % target,
+            main_src = "src/it/daml/Lib.daml",
+            srcs = glob(["src/it/daml/**/*.daml"]),
+            target = target,
+            package_prefix = "lf%s" % mangle_for_java(target)
+       )
+    ]
     for target in daml_lf_target_versions
 ]
 
@@ -110,9 +120,10 @@ daml_lf_target_versions = ["1.0", "1.1", "1.3"]
     java_test(
         name = "integration-tests-%s" % target,
         srcs = glob(["src/it/java-%s/**/*.java" % target, "src/it/java/**/*.java"]),
-        test_class = "com.digitalasset.AllTests",
+        test_class = "com.digitalasset.lf%s.AllTests" % mangle_for_java(target),
         deps = [
             ":integration-tests-model-%s.jar" % target,
+            ":integration-tests-model-noprefix-%s.jar" % target,
             "//3rdparty/jvm/org/junit/jupiter:junit_jupiter_api",
             "//3rdparty/jvm/org/junit/jupiter:junit_jupiter_engine",
             "//3rdparty/jvm/org/junit/platform:junit_platform_runner",
