+++ language-support/java/codegen/src/it/java-1.0/com/digitalasset/lf1_0/AllTests.java
@@ -1,8 +1,9 @@
 // Copyright (c) 2019 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset;
+package com.digitalasset.lf1_0;
 
+import com.digitalasset.AllGenericTests;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
